Extract TeleopNode implementation into separate module.
